moved all web stuff to web folder
