+++ web/SymAnimate.h
@@ -6,12 +6,12 @@
 #include <optional>
 #include <string>
 #include "uitsl/fetch/autoinstall.hpp"
-#include "default_mode/SymWorld.h"
-#include "default_mode/DataNodes.h"
-#include "ConfigSetup.h"
+#include "../source/default_mode/SymWorld.h"
+#include "../source/default_mode/DataNodes.h"
+#include "../source/ConfigSetup.h"
 //#include "SymJS.h"
-#include "default_mode/Symbiont.h"
-#include "default_mode/Host.h"
+#include "../source/default_mode/Symbiont.h"
+#include "../source/default_mode/Host.h"
 #include "emp/base/errors.hpp"
 #include "emp/web/Document.hpp"
 #include "emp/web/Canvas.hpp"
@@ -19,7 +19,7 @@
 #include "emp/config/ArgManager.hpp"
 #include "emp/prefab/ConfigPanel.hpp"
 #include "emp/web/UrlParams.hpp"
-#include "default_mode/WorldSetup.cc"
+#include "../source/default_mode/WorldSetup.cc"
 
 namespace UI = emp::web;
 SymConfigBase config; // load the default configuration
